Replace build command with bundle command
